added api module with Auth class and also hide modules: alchemy, base, temporary_model
